requirements.txt
numpy
pandas
matplotlib
scikit-learn
fastapi
uvicorn
streamlit
requests
pillow
